requirements.txt
Babel==2.8.0
bandit==1.6.2
blinker==1.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gitdb2==2.0.6
GitPython==3.0.5
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
passlib==1.7.2
pbr==5.4.4
pytz==2019.3
PyYAML==5.1.2
six==1.13.0
smmap2==2.0.5
speaklater==1.3
SQLAlchemy==1.3.12
stevedore==1.31.0
Werkzeug==0.16.0
WTForms==2.2.1
